refactor: split copilot github api flow
