+++ docs/CODEBASE_AUDIT_ACTION_TABLE_2026-03-09.md
@@ -62,6 +62,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | R42 | Fixed | Provider display-info split and shared fallback metric helpers | `internal/tui/model.go`, `internal/tui/model_display_info.go`, `internal/core/dashboard_display_metrics.go` | Provider tile display-summary logic moved out of the main TUI model file, and fallback/rate-limit metric selection now lives in shared core helpers instead of ad hoc TUI parsing. | Continue moving the remaining analytics/detail-specific metric decoding into shared extractors. |
 | R43 | Fixed | Codex live/session split | `internal/providers/codex/codex.go`, `internal/providers/codex/live_usage.go`, `internal/providers/codex/session_usage.go` | Codex now keeps provider wiring in the main file while live usage fetching and local session projection live in dedicated helpers. | Continue the same concern-based split for the remaining large providers. |
 | R44 | Fixed | Claude Code local file/helper split and settings modal layout split | `internal/providers/claude_code/claude_code.go`, `internal/providers/claude_code/local_files.go`, `internal/providers/claude_code/local_helpers.go`, `internal/tui/settings_modal.go`, `internal/tui/settings_modal_layout.go` | Claude Code local readers and generic helper logic are split out of the main provider file, and the settings modal layout/render wrapper no longer lives inline with all modal state/input handling. | Continue with deeper conversation-aggregation extraction in Claude Code and more TUI render-section splits. |
+| R45 | Fixed | Copilot GitHub API split | `internal/providers/copilot/copilot.go`, `internal/providers/copilot/api_data.go` | Copilot's GitHub API fetch, quota projection, and org metrics flow now live in a dedicated file instead of sharing the same unit as local config/log/session parsing. | Continue splitting the remaining local projection/helpers out of the main provider file. |
 
 ## Action Table
 
@@ -70,7 +71,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | A1 | P2 | Account config contract hardening | `internal/core/provider.go:31-43`, `internal/config/config.go:199-206` | Path overload dependence is removed from the hot runtime flow, but `Binary` / `BaseURL` still coexist in the same type and the distinction between CLI path vs provider-local path is still not encoded by type. | Introduce a dedicated typed runtime-hints/path struct and eventually retire path-related legacy comments/compatibility in `AccountConfig`. | Finishes the contract cleanup and makes misuse harder. |
 | A2 | P2 | TUI/application decomposition follow-through | `internal/tui/model.go`, `internal/tui/model_display_info.go`, `internal/tui/detail.go`, `internal/tui/detail_tokens.go`, `internal/tui/settings_modal.go`, `internal/tui/settings_modal_layout.go`, `internal/dashboardapp/service.go` | Side effects are injected, provider display-info logic is split out, and settings modal layout is separated, but TUI state-transition and render-heavy flows are still concentrated in a few large files. | Continue decomposing render-heavy/detail/settings flows and move more orchestration decisions into the dashboard application layer over time. | Lower UI complexity and smaller blast radius per change. |
 | A3 | P2 | UI metric-prefix parsing | `internal/tui/analytics.go`, `internal/tui/detail.go`, `internal/core/usage_breakdowns.go`, `internal/core/analytics_snapshot.go`, `internal/core/analytics_costs.go`, `internal/core/dashboard_display_metrics.go` | Composition bars, provider tile fallback/rate-limit selection, analytics model views, and analytics cost fallback now consume shared extractors, but some analytics/detail sections still decode metric-key conventions directly in TUI code. | Continue promoting the remaining analytics/detail extractors into `internal/core` or `internal/telemetry` so renderers consume structured sections instead of re-parsing maps. | Removes a large class of UI drift bugs and reduces per-render work. |
-| A4 | P2 | Large provider monolith follow-through | `internal/providers/ollama/ollama.go`, `internal/providers/zai/zai.go`, `internal/providers/gemini_cli/gemini_cli.go`, `internal/providers/copilot/copilot.go`, `internal/providers/claude_code/claude_code.go` | Cursor, OpenRouter, and Codex are now materially decomposed, and Claude Code has local-reader/helper splits, but several providers still combine transport, parsing, normalization, and projection in single very large files. | Split the remaining large providers by concern: account/API fetch, local-data adapters, projection helpers, telemetry helpers, and Claude Code conversation aggregation. | Smaller diffs, less drift risk, and easier provider-specific testing. |
+| A4 | P2 | Large provider monolith follow-through | `internal/providers/ollama/ollama.go`, `internal/providers/zai/zai.go`, `internal/providers/gemini_cli/gemini_cli.go`, `internal/providers/copilot/copilot.go`, `internal/providers/copilot/api_data.go`, `internal/providers/claude_code/claude_code.go` | Cursor, OpenRouter, Codex, and Copilot's API path are materially decomposed, and Claude Code has local-reader/helper splits, but several providers still combine large local parsing/projection flows in very large files. | Split the remaining large providers by concern: account/API fetch, local-data adapters, projection helpers, telemetry helpers, and Claude Code/Copilot local aggregation paths. | Smaller diffs, less drift risk, and easier provider-specific testing. |
 | A6 | P2 | Telemetry usage-view monolith | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_helpers.go`, `internal/telemetry/usage_view_projection.go`, `internal/telemetry/usage_view_queries.go`, `internal/telemetry/usage_view_materialize.go`, `internal/telemetry/usage_view_aggregate.go` | The usage-view code is materially smaller after the helper/projection/query/materialization/aggregate splits, but the top-level orchestration path still coordinates caching, source selection, and final snapshot application in one place. | Continue splitting only if future telemetry work reintroduces sprawl, and consider a typed intermediate aggregation model if query optimization pressure grows. | Easier optimization and safer incremental changes. |
 | A8 | P3 | Ambiguous local-source account attribution still requires explicit disambiguation | `internal/daemon/source_collectors.go`, `internal/daemon/server_http.go`, `cmd/openusage/telemetry.go` | Unambiguous local collectors now bind to configured accounts, but when multiple accounts share the same source paths the daemon intentionally falls back to source-scoped attribution rather than guessing. This is correct, but it still leaves ambiguous setups dependent on explicit account selection. | If multi-account local-source workflows become common, add persisted per-source alias mapping or require explicit source/account binding in config for ambiguous path groups. | Makes the remaining ambiguity explicit instead of silent, and defines the next hardening step only if needed. |
 | A7 | P3 | Daemon service follow-through | `internal/daemon/server.go`, `internal/daemon/server_collect.go`, `internal/daemon/server_spool.go`, `internal/daemon/server_poll.go`, `internal/daemon/server_http.go`, `internal/daemon/server_read_model.go` | The loop families are now separated, but the daemon still has further optimization and worker-boundary cleanup opportunities rather than a hard responsibility bug. | Keep future daemon work inside the split family files and only add a worker abstraction if concurrency pressure justifies it. | Lower mental load and easier concurrency review. |
